feat: update Metric Bot
