+++ botfarm/src/main/java/io/deeplay/camp/botfarm/Main.java
@@ -41,8 +41,8 @@ public class Main {
   }
 
   public static void BotFightFun() throws IOException {
-    ModClastMCCacheExpMaxBot bot1 = new ModClastMCCacheExpMaxBot(PlayerType.FIRST_PLAYER, 4);
-    ModNewClastMetricPlaceExpMaxBot bot2 = new ModNewClastMetricPlaceExpMaxBot(PlayerType.SECOND_PLAYER, 4);
+    ModNewClastMetricPlaceExpMaxBot bot1 = new ModNewClastMetricPlaceExpMaxBot(PlayerType.FIRST_PLAYER, 5);
+    ModNewClastMCCacheExpMaxBot bot2 = new ModNewClastMCCacheExpMaxBot(PlayerType.SECOND_PLAYER, 4);
 
     for(int i = 0; i<1;i++){
       BotFight fight = new BotFight(bot1, bot2, 100, true);
