+++ botfarm/src/main/java/io/deeplay/camp/botfarm/bots/denis_bots/UtilityFunction.java
@@ -1,8 +1,11 @@
 package io.deeplay.camp.botfarm.bots.denis_bots;
 
+import io.deeplay.camp.game.events.MakeMoveEvent;
 import io.deeplay.camp.game.events.PlaceUnitEvent;
 import io.deeplay.camp.game.mechanics.GameState;
 import io.deeplay.camp.game.mechanics.PlayerType;
+
+import java.util.List;
 
 public interface UtilityFunction {
     /**
@@ -21,4 +24,5 @@ public interface UtilityFunction {
     void setCurrentPlayerType(PlayerType playerType);
     PlayerType getCurrentPlayerType();
     double monteCarloAlg(GameState root, int countGame, PlaceUnitEvent placeUnitEvent);
+    List<MakeMoveEvent> changeMoveByTactic(GameState gameState, List<MakeMoveEvent> eventList);
 }
